# docs/toeic-exams.openapi.yaml
openapi: 3.0.0
info:
  title: Anish TOEIC Web Services
  version: 1.0.0
  description: Core API contract for the /thi-thu feature covering catalog, auth, attempts, grading, results and review.
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  example: supersecret1
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: New User
      responses:
        '201':
          description: Created — returns token and user object
          headers:
            Set-Cookie:
              schema:
                type: string
                description: httpOnly token cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                      displayName:
                        type: string
        '400':
          description: Validation error
        '409':
          description: Email already registered
  /api/auth/login:
    post:
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        '200':
          description: OK — returns token and user object
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '401':
          description: Invalid email or password
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
...